Update charts and docs (#11)
Update Charts and Documentation

Bumps the CAPI/CAPO versions to the latest point releases, fixes some typos and bugs in the documentation and updates versions as an interim fix before we have a proper solution in place to keep them in-sync with the source of truth e.g. the chart version.